♻️ refactor: move createBizOpenAI
