+++ src/app/(backend)/webapi/stt/openai/route.ts
@@ -1,7 +1,7 @@
 import { OpenAISTTPayload } from '@lobehub/tts';
 import { createOpenaiAudioTranscriptions } from '@lobehub/tts/server';
 
-import { createBizOpenAI } from '@/app/(backend)/api/openai/createBizOpenAI';
+import { createBizOpenAI } from '@/app/(backend)/_deprecated/createBizOpenAI';
 
 export const runtime = 'edge';
