+++ src/app/(backend)/webapi/tts/openai/route.ts
@@ -1,7 +1,7 @@
 import { OpenAITTSPayload } from '@lobehub/tts';
 import { createOpenaiAudioSpeech } from '@lobehub/tts/server';
 
-import { createBizOpenAI } from '@/app/(backend)/api/openai/createBizOpenAI';
+import { createBizOpenAI } from '@/app/(backend)/_deprecated/createBizOpenAI';
 
 export const runtime = 'edge';
